Fixed bug in Reader.close that occured when closed multiple times.
In Reader.close the reader object deletes its obj_locators member as
these object locators are useless without open file descriptor.
This would lead to an AttributeError if one tried to close the reader
twice as the second time this member wound't exist.
Wrapped deletion in AttributeError in a try block and now ignore the case
where this variable doesn't exist.