Added unlink_carefully method to hlab.pathutils.FilePath
This method attempts to unlink the file and in the event of an OSError
inspects errno.  If errno is ENOENT, indicating file doesn't exist,
error is ignored.  Otherwise the exception is reraised or a message is
printed (explaining file couldn't be unlinked) and this behvaior is
controlled by raise_exc parameter to unlink_carefully, which defaults
to False.